code refactoring - moving things out of the grasp planner to make sure we can test more stuff in simulation
