+++ projects/stretch_ovmm/tests/test_pick.py
@@ -1,5 +1,4 @@
-#!/usr/bin/env python
-# Copyright (c) Meta Platforms, Inc. and affiliates.
+#!/usr/birobot# Copyright (c) Meta Platforms, Inc. and affiliates.
 #
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
@@ -73,7 +72,7 @@ def run_experiment(visualize_maps=False, test_id=0, reset_nav=False, **kwargs):
         robot.nav.navigate_to([0, 0, 0])
 
     # Put it into initial posture
-    env.grasp_planner.go_to_manip_mode()
+    env.robot.go_to_manip_mode()
 
     # do some tests
     if test_id == 0:
